# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Fishcake ecosystem - activity, drop, staking, mining and leaderboard",
        "title": "Fishcake Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/v1/activity/info": {
            "get": {
                "description": "Get single activity info by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity"
                ],
                "summary": "Activity detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Activity ID",
                        "name": "activityId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_result.ApiResult"
                        }
                    }
                }
            }
        },
        "/v1/activity/list": {
            "get": {
                "description": "Query activity list with filters, corresponds to contract ActivityAdd events",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity"
                ],
                "summary": "Activity list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number, starts from 1",
                        "name": "pageNum",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Activity ID",
                        "name": "activityId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status: 1=active, 2=finished, 3=expired",
                        "name": "activityStatus",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Business name (fuzzy search)",
                        "name": "businessName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Business wallet address (fuzzy search)",
                        "name": "businessAccount",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Token contract address (fuzzy search)",
                        "name": "tokenContractAddr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Latitude (requires longitude and scope)",
                        "name": "latitude",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Longitude",
                        "name": "longitude",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search radius in meters",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "NFT filter: 1=Pro NFT valid, 2=Basic NFT valid",
                        "name": "activityFilter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_result.ApiResult"
                        }
                    }
                }
            }
        },
        "/v1/activity/minedAmount": {
            "get": {
                "description": "Get cumulative mining amount for a user (business account)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity"
                ],
                "summary": "Get user mined amount",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "1=last 30 days only, omit for all-time",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api_result.ApiResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "obj": {
                                            "type": "object",
                                            "properties": {
                                                "month": {
                                                    "type": "boolean"
                                                },
                                                "totalMined": {
                                                    "type": "string"
                                                },
                                                "userAddress": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/activity/miningInfo": {
            "get": {
                "description": "Get real-time mining info including Fishcake Power for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Earnings"
                ],
                "summary": "User mining info \u0026 Fishcake Power",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_result.ApiResult"
                        }
                    }
                }
            }
        },
        "/v1/activity/miningRank": {
            "get": {
                "description": "Get mining rank by business account mined amount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leaderboard"
                ],
                "summary": "Mining leaderboard",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "1=last 30 days, omit for all-time",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_result.ApiResult"
                        }
                    }
                }
            }
        },
...
        "/v1/drop/list": {
            "get": {
                "description": "Query drop/receive records, corresponds to contract Drop events",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History"
                ],
                "summary": "Drop/Receive records",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number, starts from 1",
                        "name": "pageNum",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Wallet address (case insensitive)",
                        "name": "address",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "1=user received, 2=merchant dropped",
                        "name": "dropType",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_result.ApiResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api_result.ApiResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "obj": {},
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}